compelete axios, myprofile page, rowitem page, login, register, home page, moresong page, person page,persons page ,header, add some new functions

+++ axios/axios.js
@@ -1,17 +1,26 @@
 import axios from "axios";
 let tryCount = 0;
+let downloaderTryCount = 0;
 
 const instanceApi = axios.create({
   baseURL: "https://nejat.safine.co/api",
 });
+// instanceApi.interceptors.request.use((request) => {
+//   console.log(request);
+
+//   return request;
+// });
+
 instanceApi.interceptors.response.use(
   (response) => {
+    // console.log(response);
+
     return response;
   },
 
   async (error) => {
     console.log(error.config);
-    // console.log(error.response);https://nejat.safine.co/api
+    // console.log(error.response);
     // const { logout } = useContext(authContext);
     // console.log(tryCount);
     // console.log(
@@ -21,19 +30,27 @@ instanceApi.interceptors.response.use(
     //   error.response.status,
     //   error.config
     // );
-    // if (error.config && error.response && error.response.status === 404) {
-    //   window.location = '/aboutus';
-    //   return Promise.reject(error);
-    // }
+    if (error.config && error.response && error.response.status === 404) {
+      window.location = "/not_found";
+      return Promise.reject(error);
+      // console.log(111);
+    }
 
     if (error.config && error.response && error.response.status === 401) {
       if (tryCount === 3) {
         // console.log(66);
-        localStorage.clear();
-        window.location = "/login";
+        const res = JSON.parse(
+          localStorage.getItem("logForRefreshTokenExpired")
+        );
+        if (!res) {
+          localStorage.clear();
+          window.location = "/login";
+          localStorage.setItem("logForRefreshTokenExpired", true);
+        }
+
         return Promise.reject(error);
       }
-      // console.log(error.config);
+      // console.log(tryCount);
       // console.log(error.response && error.response.status === 401);
       const config = {
         headers: {
@@ -69,16 +86,76 @@ instanceApi.interceptors.response.use(
 );
 
 const downloader = axios.create({
+  // baseURL: "http://downloader.7negare.ir/download",
   baseURL: "https://downloader.safine.co/",
 });
-// cancelTokenSource.cancel();
+// downloader.interceptors.request.use((request) => {
+//   console.log(request);
 
-// const auth = axios.create({
-//   baseURL: 'https://nejat.safine.co/api/account',
+//   return request;
 // });
+
+downloader.interceptors.response.use(
+  (response) => {
+    // console.log(response);
+    return response;
+  },
+
+  async (error) => {
+    console.log(error.config);
+
+    if (error.config && error.response && error.response.status === 503) {
+      if (downloaderTryCount === 3) {
+        return Promise.reject(error);
+      }
+
+      downloaderTryCount++;
+
+      return setTimeout(() => {
+        instanceApi.request(error.config);
+      }, 3000);
+    }
+
+    return Promise.reject(error);
+  }
+);
+
+// for getting photos
+const simpleApi = axios.create({
+  baseURL: "https://nejat.safine.co/api",
+});
+simpleApi.interceptors.response.use(
+  (response) => {
+    // console.log(response);
+
+    return response;
+  },
+
+  async (error) => {
+    // console.log(error.config);
+    // console.log(error.response);
+    // const { logout } = useContext(authContext);
+    // console.log(tryCount);
+    // console.log(
+    //   '**t***',
+    //   error.config && error.response && error.response.status === 401,
+    //   error,
+    //   error.response.status,
+    //   error.config
+    // );
+    if (error.config && error.response && error.response.status === 404) {
+      window.location = "/404";
+      return Promise.reject(error);
+      // console.log(111);
+    }
+
+    return Promise.reject(error);
+  }
+);
 // eslint-disable-next-line
 export default {
   instanceApi,
   downloader,
+  simpleApi,
   // , auth
 };
